Cleaning up code
Encrypter.py:

Cleaned up the class methods a bit to increase readability and make it more modular.

Added a new argument "stdin" to help facilitate outputs.

binlock.py:

Moved checking and handling of stdin to the main file.

Changed the format of the Argument printing section and added a debug switch to access that code without manually un-commenting it, and removed the -S stdout print, as --debug does what that does.

Updated README.me and took TODO out of binlock.py

Updated the dev-test script with a sleep for no reason.