v1.04 The Big AES Patch
binlock now supports encrypting files with AES instead of just encoding. Currently you can only use a password, but I will add more functionality eventually.

for the main program other than AES I've made some slight format changes to the code, changed the arguments and help a bit. Updated the license with the new program description

encoder.py is now bintools.py since it does more than just encoding now. we have a ton of new functions in there for you to look at.

made a few small changes to the way build.sh works. I need to check for dependencies eventually. You need snapcraft, multipass, zip and probably some other things. But for now we'll leave it the way it is.

I removed dev-test.sh since the program is kinda out of scope for it. The script still works for encoding, but it can't test the encryption stuff yet.

updated README.me to include the new description and a few small changes.